move runtime-interface tests
