+++ primitives/runtime-interface/src/lib.rs
@@ -117,96 +117,3 @@ pub type Pointer<T> = *mut T;
 /// A pointer that can be used in a runtime interface function signature.
 #[cfg(feature = "std")]
 pub type Pointer<T> = wasm_interface::Pointer<T>;
-
-#[cfg(test)]
-mod tests {
-	use super::*;
-	use test_wasm::{WASM_BINARY, test_api::HostFunctions};
-	use wasm_interface::HostFunctions as HostFunctionsT;
-
-	type TestExternalities = state_machine::TestExternalities<primitives::Blake2Hasher, u64>;
-
-	fn call_wasm_method<HF: HostFunctionsT>(method: &str) -> TestExternalities {
-		let mut ext = TestExternalities::default();
-		let mut ext_ext = ext.ext();
-
-		executor::call_in_wasm::<
-			_,
-			(
-				HF,
-				sp_io::SubstrateHostFunctions,
-				executor::deprecated_host_interface::SubstrateExternals
-			)
-		>(
-			method,
-			&[],
-			executor::WasmExecutionMethod::Interpreted,
-			&mut ext_ext,
-			&WASM_BINARY[..],
-			8,
-		).expect(&format!("Executes `{}`", method));
-
-		ext
-	}
-
-	#[test]
-	fn test_return_data() {
-		call_wasm_method::<HostFunctions>("test_return_data");
-	}
-
-	#[test]
-	fn test_return_option_data() {
-		call_wasm_method::<HostFunctions>("test_return_option_data");
-	}
-
-	#[test]
-	fn test_set_storage() {
-		let mut ext = call_wasm_method::<HostFunctions>("test_set_storage");
-
-		let expected = "world";
-		assert_eq!(expected.as_bytes(), &ext.ext().storage("hello".as_bytes()).unwrap()[..]);
-	}
-
-	#[test]
-	fn test_return_value_into_mutable_reference() {
-		call_wasm_method::<HostFunctions>("test_return_value_into_mutable_reference");
-	}
-
-	#[test]
-	fn test_get_and_return_array() {
-		call_wasm_method::<HostFunctions>("test_get_and_return_array");
-	}
-
-	#[test]
-	fn test_array_as_mutable_reference() {
-		call_wasm_method::<HostFunctions>("test_array_as_mutable_reference");
-	}
-
-	#[test]
-	fn test_return_input_public_key() {
-		call_wasm_method::<HostFunctions>("test_return_input_public_key");
-	}
-
-	#[test]
-	#[should_panic(
-		expected = "Other(\"Instantiation: Export ext_test_api_return_input_version_1 not found\")"
-	)]
-	fn host_function_not_found() {
-		call_wasm_method::<()>("test_return_data");
-	}
-
-	#[test]
-	#[should_panic(
-		expected =
-			"FunctionExecution(\"ext_test_api_invalid_utf8_data_version_1\", \
-			\"Invalid utf8 data provided\")"
-	)]
-	fn test_invalid_utf8_data_should_return_an_error() {
-		call_wasm_method::<HostFunctions>("test_invalid_utf8_data_should_return_an_error");
-	}
-
-	#[test]
-	fn test_overwrite_native_function_implementation() {
-		call_wasm_method::<HostFunctions>("test_overwrite_native_function_implementation");
-	}
-}
